made some changes to resolve some issues
