+++ src/components/table/data-table-multi-select-actions/trace-table-multi-select-action.tsx
@@ -1,5 +1,4 @@
-/* eslint-disable */
-// @ts-nocheck
+import { ChevronDown } from "lucide-react";
 
 import {
   DropdownMenu,
@@ -7,19 +6,26 @@ import {
   DropdownMenuContent,
   DropdownMenuItem,
 } from "@/src/components/ui/dropdown-menu";
-import { DeleteTraceMultiSelectAction } from "@/src/components/delete-trace";
 import { Button } from "@/src/components/ui/button";
-import { ChevronDown } from "lucide-react";
+import { useHasAccess } from "@/src/features/rbac/utils/checkAccess";
+import { api } from "@/src/utils/api";
 
 export function TraceTableMultiSelectAction({
   selectedRows,
   projectId,
 }: {
-  selectedRows: object[];
+  selectedRows: any[];
   projectId: string;
 }) {
+  const utils = api.useUtils();
+
+  const hasAccess = useHasAccess({ projectId, scope: "traces:delete" });
+
+  const mutDeleteTraces = api.traces.deleteMany.useMutation({
+    onSuccess: () => void utils.traces.invalidate(),
+  });
+
   const traceIds = selectedRows.map((row) => {
-    // @ts-ignore
     return row.original.id;
   });
 
@@ -31,16 +37,16 @@ export function TraceTableMultiSelectAction({
           className="bg-white p-2 font-medium text-black"
           disabled={selectedRows.length < 1}
         >
-          Actions
+          Actions ({selectedRows.length} selected)
           <ChevronDown className="h-5 w-5" />
         </Button>
       </DropdownMenuTrigger>
-      <DropdownMenuContent className="text-center">
-        <DropdownMenuItem>
-          <DeleteTraceMultiSelectAction
-            traceIds={traceIds}
-            projectId={projectId}
-          />
+      <DropdownMenuContent>
+        <DropdownMenuItem
+          disabled={!hasAccess}
+          onClick={() => void mutDeleteTraces.mutateAsync({ traceIds, projectId })}
+        >
+          Delete
         </DropdownMenuItem>
       </DropdownMenuContent>
     </DropdownMenu>
